needed to expose the match class for feature work
